js lib update
