+++ web_interface/back_front/model_blocks.py
@@ -257,7 +257,8 @@ class ModelManagerBlock(Block):
         assert self.gnn is not None
 
         # Import correct class
-        from web_interface.main_multi import FrontendClient
+        # from web_interface.main_multi import FrontendClient
+        from web_interface.main_fastapi import FrontendClient
         if self.klass in FrontendClient.get_parameters("FW"):
             mm_class = import_by_name(self.klass, ["models_builder.gnn_models"])
