+++ web_interface/static/js/utils.js
@@ -399,6 +399,7 @@ function sleep(ms) {
 }
 
 function JSON_parse(string) {
+    // console.log('parsing JSON from string:', string)
     return JSON.parse(string, function (key, value) {
         if (value === 'NaN')
             return NaN
