+++ web_interface/templates/base.html
@@ -126,7 +126,7 @@
     </span>
   </div>
 
-  <script src="static/js/libs/socket.io-4.0.1.js"></script>
+  <script src="static/js/libs/socket.io-4.6.1.js"></script>
   <script src="static/js/libs/jquery-3.6.0.js"></script>
   <script src="static/js/libs/chart-3.7.1.min.js"></script>
   <script src="static/js/libs/datatables-1.13.2.min.js"></script>
